feat: Update Route Model and Populate Stations
- Added migration to update Route model, renaming TrainId to StationId and adding IntermediateStations and Train properties.
- Created migration to populate Stations table with initial data.
- Implemented views for adding and managing routes, including form validation and data binding for trains and stations.
- Updated response headers in static web assets for consistency.
- Updated sourcelink and PDB files for debugging purposes.

+++ Services/RouteService.cs
@@ -85,8 +85,9 @@ public class RouteService
                 graph[route.Destination] = new Dictionary<string, int>();
             }
 
-            graph[route.Origin][route.Destination] = route.Distance;
-            graph[route.Destination][route.Origin] = route.Distance; // Grafo no dirigido
+            // Conversión explícita de double a int
+            graph[route.Origin][route.Destination] = (int)route.Distance;
+            graph[route.Destination][route.Origin] = (int)route.Distance; // Grafo no dirigido
         }
 
         return graph;
